# swagger.yml
swagger: '2.0'
info:
  title: Gwent API
  description: API браузерной мультиплеерной карточной игры ТехноГвинт
  version: "1.0.0"
  contact:
    email: kgulyy@my.com
  license:
    name: MIT License
    url: https://github.com/java-park-mail-ru/Gwent-09-2017/blob/master/LICENSE.md
schemes:
- https
host: 'tech-gwent.xyz'
basePath: /api
consumes:
- application/json
produces:
- application/json
paths:
...
  /users:
    get:
      summary: Получение списка пользователей
      description: |
        Получение данных о пользователях, отсортированные по убыванию кол-ва побед и возрастанию логина
      operationId: getUsersOnScoreBoard
      parameters:
      - name: Origin
        in: header
        description: Домен, с которого осуществлён запрос.
        required: true
        type: string
      - name: limit
        in: query
        type: number
        default: 10
        minimum: 0
        description: Максимальное кол-во возвращаемых записей.
      - name: offset
        in: query
        type: number
        default: 0
        description: Позиция с которой возвращаются записи.
      responses:
        200:
          description: |
            Данные авторизованного пользователя.
          schema:
            $ref: '#/definitions/Users'
        401:
          description: |
            Пользователь не авторизован.
          schema:
            $ref: '#/definitions/Error'
definitions:
...
  UserInfo:
    description: Информация о пользователе.
    type: object
    properties:
      login:
        type: string
        format: identity
        description: Имя пользователя (уникальное поле).
        example: captain
        x-isnullable: false
      email:
        type: string
        format: email
        description: Почтовый адрес пользователя.
        example: captain@blackpearl.sea
        x-isnullable: true
      wins:
        type: integer
        description: Кол-во выигранных партий
        example: 5
        x-isnullable: false
      position:
        type: integer
        description: Позиция в таблице лидеров
        example: 8
        x-isnullable: true
  Users:
    type: array
    items:
      $ref: '#/definitions/UserInfo'
...
  Error:
    type: object
    properties:
      message:
        type: string
        readOnly: true
        description: Текстовое описание ошибки.
        example: Error message example